working on refresh
