+++ app/src/main/res/layout/activity_main.xml
@@ -42,17 +42,11 @@
 		    app:tabMode="scrollable" />
     </android.support.design.widget.AppBarLayout>
 
-	<android.support.v4.widget.SwipeRefreshLayout
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:id="@+id/swipeRefreshLayout">
-
 		<android.support.v4.view.ViewPager
 			android:id="@+id/viewpager"
 			android:layout_width="match_parent"
 			android:layout_height="match_parent"
 			app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-	</android.support.v4.widget.SwipeRefreshLayout>
 
 	<ProgressBar
 		android:id="@+id/loadingBar"
